Fix C interface to match Senzing encryption plugin spec
- Return type: int (i32) -> int64_t (i64) on all 8 exported functions
- Error codes: use spec values (0, -1, -5, -20, -30) instead of
  invented codes (-2 through -7, -99)
- InitPlugin param: void* -> const struct CParameterList* per spec
- Add CParameterList/CParameterTuple repr(C) structs
- Add error code constants matching g2EncryptionPluginInterface_defs.h
- Update header, README, and CHANGELOG accordingly

Ref: https://github.com/Senzing/senzing-data-encryption-specification

+++ sz_common/src/utils.rs
@@ -114,7 +114,7 @@ pub unsafe fn error_to_c_buffer(
     error_buffer: *mut libc::c_char,
     max_error_size: usize,
     error_size: *mut usize,
-) -> i32 {
+) -> i64 {
     let error_code = error.to_error_code();
     let error_message = error.to_string();
